// FudNow/src/app/dashboard/model/menu-item.ts
export class MenuItem {
    isMenuCategory: boolean;
    categoryName: string;
    id: number;
    name: string;
    type: string;
    description: string;
    price: number;
    itemCategory: number;
    itemCount: number;
    rating: number;
    veg: boolean;
    spicy: boolean;
    healthy: boolean;
}